better code structure but still a lot to do here.
