+++ cryptoAPI/src/asymmetric.py
@@ -1,5 +1,5 @@
 from cryptography.hazmat.primitives.asymmetric import rsa
 
 
-class Asymetric:
+class Asymmetric:
     pass
