--- cryptoAPI/src/symetric.py
@@ -1,6 +1,0 @@
-from crypto import Crypto
-import base64
-from cryptography.fernet import Fernet
-
-
-class Symetric:
